# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: UEMF University Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /faculty:
    get:
      operationId: listFaculty
      tags: [faculty]
      summary: List all faculty members
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: departmentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of faculty
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FacultyMember"
    post:
      operationId: createFaculty
      tags: [faculty]
      summary: Create a faculty member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFacultyBody"
      responses:
        "201":
          description: Created faculty member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacultyMember"
  /faculty/{id}:
    get:
      operationId: getFaculty
      tags: [faculty]
      summary: Get a faculty member by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Faculty member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacultyMember"
    put:
      operationId: updateFaculty
      tags: [faculty]
      summary: Update a faculty member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFacultyBody"
      responses:
        "200":
          description: Updated faculty member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacultyMember"
    delete:
      operationId: deleteFaculty
      tags: [faculty]
      summary: Delete a faculty member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── DEPARTMENTS ──────────────────────────────────────────────
...
  /grades:
    get:
      operationId: listGrades
      tags: [grades]
      summary: List grades
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
        - name: courseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of grades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grade"
    post:
      operationId: createGrade
      tags: [grades]
      summary: Record a grade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGradeBody"
      responses:
        "201":
          description: Created grade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grade"
  /grades/{id}:
    put:
      operationId: updateGrade
      tags: [grades]
      summary: Update a grade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGradeBody"
      responses:
        "200":
          description: Updated grade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grade"
    delete:
      operationId: deleteGrade
      tags: [grades]
      summary: Delete a grade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SCHEDULES ────────────────────────────────────────────────
  /schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      summary: List class schedules
      parameters:
        - name: courseId
          in: query
          schema:
            type: integer
        - name: facultyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
    post:
      operationId: createSchedule
      tags: [schedules]
      summary: Create a schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleBody"
      responses:
        "201":
          description: Created schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
  /schedules/{id}:
    put:
      operationId: updateSchedule
      tags: [schedules]
      summary: Update a schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleBody"
      responses:
        "200":
          description: Updated schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      summary: Delete a schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── EXAMS ────────────────────────────────────────────────────
...
  /fees:
    get:
      operationId: listFees
      tags: [fees]
      summary: List fee records
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, overdue, waived]
      responses:
        "200":
          description: List of fees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fee"
    post:
      operationId: createFee
      tags: [fees]
      summary: Create a fee record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeeBody"
      responses:
        "201":
          description: Created fee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fee"
  /fees/{id}:
    put:
      operationId: updateFee
      tags: [fees]
      summary: Update a fee record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeeBody"
      responses:
        "200":
          description: Updated fee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fee"
...
components:
  schemas:
...
    FacultyMember:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        departmentId:
          type: integer
        departmentName:
          type: string
        title:
          type: string
          enum: [professor, associate_professor, assistant_professor, lecturer, instructor]
        specialization:
          type: string
        officeLocation:
          type: string
        officeHours:
          type: string
        courseCount:
          type: integer
        status:
          type: string
          enum: [active, inactive, on_leave]
        createdAt:
          type: string
      required:
        - id
        - employeeId
        - firstName
        - lastName
        - email
        - departmentId
        - title
        - status
    CreateFacultyBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        departmentId:
          type: integer
        title:
          type: string
          enum: [professor, associate_professor, assistant_professor, lecturer, instructor]
        specialization:
          type: string
        officeLocation:
          type: string
        officeHours:
          type: string
        status:
          type: string
          enum: [active, inactive, on_leave]
      required: [firstName, lastName, email, departmentId, title, status]

    # ─── Department ──────────────────────────────────────────────
...
    Grade:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        courseId:
          type: integer
        courseName:
          type: string
        courseCode:
          type: string
        score:
          type: number
        letterGrade:
          type: string
        gradePoints:
          type: number
        semester:
          type: string
        assessmentType:
          type: string
          enum: [midterm, final, assignment, quiz, project]
        comments:
          type: string
        gradedAt:
          type: string
      required:
        - id
        - studentId
        - courseId
        - score
        - letterGrade
        - semester
        - assessmentType
    CreateGradeBody:
      type: object
      properties:
        studentId:
          type: integer
        courseId:
          type: integer
        score:
          type: number
        semester:
          type: string
        assessmentType:
          type: string
          enum: [midterm, final, assignment, quiz, project]
        comments:
          type: string
      required: [studentId, courseId, score, semester, assessmentType]

    # ─── Schedule ────────────────────────────────────────────────
    Schedule:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        courseName:
          type: string
        courseCode:
          type: string
        facultyId:
          type: integer
        facultyName:
          type: string
        dayOfWeek:
          type: string
          enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
        startTime:
          type: string
        endTime:
          type: string
        room:
          type: string
        building:
          type: string
        semester:
          type: string
        type:
          type: string
          enum: [lecture, lab, tutorial, seminar]
      required:
        - id
        - courseId
        - dayOfWeek
        - startTime
        - endTime
        - room
        - semester
        - type
    CreateScheduleBody:
      type: object
      properties:
        courseId:
          type: integer
        facultyId:
          type: integer
        dayOfWeek:
          type: string
          enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
        startTime:
          type: string
        endTime:
          type: string
        room:
          type: string
        building:
          type: string
        semester:
          type: string
        type:
          type: string
          enum: [lecture, lab, tutorial, seminar]
      required: [courseId, dayOfWeek, startTime, endTime, room, semester, type]

    # ─── Exam ────────────────────────────────────────────────────
...
    Fee:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        type:
          type: string
          enum: [tuition, registration, library, lab, dormitory, other]
        amount:
          type: number
        dueDate:
          type: string
        paidDate:
          type: string
        status:
          type: string
          enum: [pending, paid, overdue, waived]
        semester:
          type: string
        description:
          type: string
        createdAt:
          type: string
      required:
        - id
        - studentId
        - type
        - amount
        - dueDate
        - status
        - semester
    CreateFeeBody:
      type: object
      properties:
        studentId:
          type: integer
        type:
          type: string
          enum: [tuition, registration, library, lab, dormitory, other]
        amount:
          type: number
        dueDate:
          type: string
        paidDate:
          type: string
        status:
          type: string
          enum: [pending, paid, overdue, waived]
        semester:
          type: string
        description:
          type: string
      required: [studentId, type, amount, dueDate, status, semester]
...